Ingredient refactor 2
